+++ examples/provider/provider.tf
@@ -7,11 +7,11 @@ terraform {
 }
 
 provider "truenas" {
-  base_url = "http://10.0.0.213:81"
+  base_url = "http://10.0.0.83:81"
   api_key  = "your-api-key-here"
 }
 
 # Or use environment variables:
-# export TRUENAS_BASE_URL="http://10.0.0.213:81"
+# export TRUENAS_BASE_URL="http://10.0.0.83:81"
 # export TRUENAS_API_KEY="your-api-key-here"